Move numba subtensor functionality to its own module
