+++ tests/link/numba/test_basic.py
@@ -33,7 +33,6 @@ from pytensor.link.numba.linker import NumbaLinker
 from pytensor.raise_op import assert_op
 from pytensor.scalar.basic import ScalarOp, as_scalar
 from pytensor.tensor import blas
-from pytensor.tensor import subtensor as pt_subtensor
 from pytensor.tensor.elemwise import Elemwise
 from pytensor.tensor.shape import Reshape, Shape, Shape_i, SpecifyShape
 
@@ -360,218 +359,6 @@ def test_get_numba_type(v, expected, force_scalar, not_implemented):
 def test_create_numba_signature(v, expected, force_scalar):
     res = numba_basic.create_numba_signature(v, force_scalar=force_scalar)
     assert res == expected
-
-
-@pytest.mark.parametrize(
-    "x, indices",
-    [
-        (pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))), (1,)),
-        (
-            pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))),
-            (slice(None)),
-        ),
-        (pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))), (1, 2, 0)),
-        (
-            pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))),
-            (slice(1, 2), 1, slice(None)),
-        ),
-    ],
-)
-def test_Subtensor(x, indices):
-    """Test NumPy's basic indexing."""
-    out_pt = x[indices]
-    assert isinstance(out_pt.owner.op, pt_subtensor.Subtensor)
-    out_fg = FunctionGraph([], [out_pt])
-    compare_numba_and_py(out_fg, [])
-
-
-@pytest.mark.parametrize(
-    "x, indices",
-    [
-        (pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))), ([1, 2],)),
-        (pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))), ([1],)),
-    ],
-)
-def test_AdvancedSubtensor1(x, indices):
-    """Test NumPy's advanced indexing in one dimension."""
-    out_pt = pt_subtensor.advanced_subtensor1(x, *indices)
-    assert isinstance(out_pt.owner.op, pt_subtensor.AdvancedSubtensor1)
-    out_fg = FunctionGraph([], [out_pt])
-    compare_numba_and_py(out_fg, [])
-
-
-def test_AdvancedSubtensor1_out_of_bounds():
-    out_pt = pt_subtensor.advanced_subtensor1(np.arange(3), [4])
-    assert isinstance(out_pt.owner.op, pt_subtensor.AdvancedSubtensor1)
-    out_fg = FunctionGraph([], [out_pt])
-    with pytest.raises(IndexError):
-        compare_numba_and_py(out_fg, [])
-
-
-@pytest.mark.parametrize(
-    "x, indices",
-    [
-        (pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))), ([1, 2], [2, 3])),
-        (
-            pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))),
-            ([1, 2], slice(None), [3, 4]),
-        ),
-    ],
-)
-def test_AdvancedSubtensor(x, indices):
-    """Test NumPy's advanced indexing in more than one dimension."""
-    out_pt = x[indices]
-    assert isinstance(out_pt.owner.op, pt_subtensor.AdvancedSubtensor)
-    out_fg = FunctionGraph([], [out_pt])
-    compare_numba_and_py(out_fg, [])
-
-
-@pytest.mark.parametrize(
-    "x, y, indices",
-    [
-        (
-            pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))),
-            pt.as_tensor(np.array(10)),
-            (1,),
-        ),
-        (
-            pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))),
-            pt.as_tensor(rng.poisson(size=(4, 5))),
-            (slice(None)),
-        ),
-        (
-            pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))),
-            pt.as_tensor(np.array(10)),
-            (1, 2, 0),
-        ),
-        (
-            pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))),
-            pt.as_tensor(rng.poisson(size=(1, 5))),
-            (slice(1, 2), 1, slice(None)),
-        ),
-    ],
-)
-def test_IncSubtensor(x, y, indices):
-    out_pt = pt.set_subtensor(x[indices], y)
-    assert isinstance(out_pt.owner.op, pt_subtensor.IncSubtensor)
-    out_fg = FunctionGraph([], [out_pt])
-    compare_numba_and_py(out_fg, [])
-
-    out_pt = pt.inc_subtensor(x[indices], y)
-    assert isinstance(out_pt.owner.op, pt_subtensor.IncSubtensor)
-    out_fg = FunctionGraph([], [out_pt])
-    compare_numba_and_py(out_fg, [])
-
-    x_pt = x.type()
-    out_pt = pt.set_subtensor(x_pt[indices], y, inplace=True)
-    assert isinstance(out_pt.owner.op, pt_subtensor.IncSubtensor)
-    out_fg = FunctionGraph([x_pt], [out_pt])
-    compare_numba_and_py(out_fg, [x.data])
-
-
-@pytest.mark.parametrize(
-    "x, y, indices",
-    [
-        (
-            pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))),
-            pt.as_tensor(rng.poisson(size=(2, 4, 5))),
-            ([1, 2],),
-        ),
-        (
-            pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))),
-            pt.as_tensor(rng.poisson(size=(2, 4, 5))),
-            ([1, 1],),
-        ),
-        # Broadcasting values
-        (
-            pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))),
-            pt.as_tensor(rng.poisson(size=(1, 4, 5))),
-            ([0, 2, 0],),
-        ),
-        (
-            pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))),
-            pt.as_tensor(rng.poisson(size=(5,))),
-            ([0, 2],),
-        ),
-        (
-            pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))),
-            pt.as_tensor(rng.poisson(size=())),
-            ([2, 0],),
-        ),
-        (
-            pt.as_tensor(np.arange(5)),
-            pt.as_tensor(rng.poisson(size=())),
-            ([2, 0],),
-        ),
-    ],
-)
-def test_AdvancedIncSubtensor1(x, y, indices):
-    out_pt = pt_subtensor.advanced_set_subtensor1(x, y, *indices)
-    assert isinstance(out_pt.owner.op, pt_subtensor.AdvancedIncSubtensor1)
-    out_fg = FunctionGraph([], [out_pt])
-    compare_numba_and_py(out_fg, [])
-
-    out_pt = pt_subtensor.advanced_inc_subtensor1(x, y, *indices)
-    assert isinstance(out_pt.owner.op, pt_subtensor.AdvancedIncSubtensor1)
-    out_fg = FunctionGraph([], [out_pt])
-    compare_numba_and_py(out_fg, [])
-
-    # With symbolic inputs
-    x_pt = x.type()
-    y_pt = y.type()
-
-    out_pt = pt_subtensor.AdvancedIncSubtensor1(inplace=True)(x_pt, y_pt, *indices)
-    assert isinstance(out_pt.owner.op, pt_subtensor.AdvancedIncSubtensor1)
-    out_fg = FunctionGraph([x_pt, y_pt], [out_pt])
-    compare_numba_and_py(out_fg, [x.data, y.data])
-
-    out_pt = pt_subtensor.AdvancedIncSubtensor1(set_instead_of_inc=True, inplace=True)(
-        x_pt, y_pt, *indices
-    )
-    assert isinstance(out_pt.owner.op, pt_subtensor.AdvancedIncSubtensor1)
-    out_fg = FunctionGraph([x_pt, y_pt], [out_pt])
-    compare_numba_and_py(out_fg, [x.data, y.data])
-
-
-@pytest.mark.parametrize(
-    "x, y, indices",
-    [
-        (
-            pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))),
-            pt.as_tensor(rng.poisson(size=(2, 5))),
-            ([1, 2], [2, 3]),
-        ),
-        (
-            pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))),
-            pt.as_tensor(rng.poisson(size=(2, 4))),
-            ([1, 2], slice(None), [3, 4]),
-        ),
-        pytest.param(
-            pt.as_tensor(np.arange(3 * 4 * 5).reshape((3, 4, 5))),
-            pt.as_tensor(rng.poisson(size=(2, 5))),
-            ([1, 1], [2, 2]),
-        ),
-    ],
-)
-def test_AdvancedIncSubtensor(x, y, indices):
-    out_pt = pt.set_subtensor(x[indices], y)
-    assert isinstance(out_pt.owner.op, pt_subtensor.AdvancedIncSubtensor)
-    out_fg = FunctionGraph([], [out_pt])
-    compare_numba_and_py(out_fg, [])
-
-    out_pt = pt.inc_subtensor(x[indices], y)
-    assert isinstance(out_pt.owner.op, pt_subtensor.AdvancedIncSubtensor)
-    out_fg = FunctionGraph([], [out_pt])
-    compare_numba_and_py(out_fg, [])
-
-    x_pt = x.type()
-    out_pt = pt.set_subtensor(x_pt[indices], y)
-    # Inplace isn't really implemented for `AdvancedIncSubtensor`, so we just
-    # hack it on here
-    out_pt.owner.op.inplace = True
-    assert isinstance(out_pt.owner.op, pt_subtensor.AdvancedIncSubtensor)
-    out_fg = FunctionGraph([x_pt], [out_pt])
-    compare_numba_and_py(out_fg, [x.data])
 
 
 @pytest.mark.parametrize(
